pyproject.toml
[tool.poetry]
name = "crewai-28"
version = "0.28.8"
description = "AI Suppliers research system using CrewAI"
authors = ["Your Name <you@example.com>"]
packages = [
    { include = "crewai", from = "src" },
]

[tool.poetry.urls]
Homepage = "https://crewai.com"
Documentation = "https://github.com/joaomdmoura/CrewAI/wiki/Index"
Repository = "https://github.com/joaomdmoura/crewai"


[tool.poetry.dependencies]
streamlit = "^1.13.0"
pysqlite3-binary = "^0.5.1"
python = ">=3.10,<=3.13"
pydantic = "^2.4.2"
langchain = "^0.1.10"
openai = "^1.13.3"
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-exporter-otlp-proto-http = "^1.22.0"
instructor = "^0.5.2"
regex = "^2023.12.25"
crewai-tools = { version = "^0.1.7", optional = true }
click = "^8.1.7"
python-dotenv = "1.0.0"
embedchain = "^0.1.98"
appdirs = "^1.4.4"
python-decouple = "^3.6"
crewai = "^0.28.8"
# ✅ Added for production server deployment

[tool.poetry.extras]
tools = ["crewai-tools"]

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
pyright = ">=1.1.350,<2.0.0"
autoflake = "^2.2.1"
pre-commit = "^3.6.0"
mkdocs = "^1.4.3"
mkdocstrings = "^0.22.0"
mkdocstrings-python = "^1.1.2"
mkdocs-material = {extras = ["imaging"], version = "^9.5.7"}
mkdocs-material-extensions = "^1.3.1"
pillow = "^10.2.0"
cairosvg = "^2.7.1"

[tool.isort]
profile = "black"
known_first_party = ["crewai"]

[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"
pytest-vcr = "^1.0.2"
python-dotenv = "1.0.0"

[tool.poetry.scripts]
crewai = "crewai.cli.cli:crewai" # ✅ Added for production deployment

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.crewai]
type = "crew"
